+++ core/src/elements/ons-tabbar/index.spec.js
@@ -232,8 +232,8 @@ describe('OnsTabbarElement', () => {
       template = ons._util.createElement('<ons-template id="page1"><ons-page>Page1</ons-page></ons-template>');
       element = ons._util.createElement(`
         <ons-tabbar>
-          <ons-tab label="Page 1" page="page1" no-reload></ons-tab>
-          <ons-tab label="Page 2" page="page1" no-reload></ons-tab>
+          <ons-tab label="Page 1" page="page1"></ons-tab>
+          <ons-tab label="Page 2" page="page1"></ons-tab>
         </ons-tabbar>
       `);
       document.body.appendChild(template);
@@ -266,18 +266,24 @@ describe('OnsTabbarElement', () => {
       let promise = new Promise((resolve) => {
         element.addEventListener('hide', resolve);
       });
-      element.setActiveTab(0);
-      element.setActiveTab(1);
-      return expect(promise).to.eventually.be.fulfilled;
+
+      return element.setActiveTab(0).then(function() {
+        element.setActiveTab(1);
+        return expect(promise).to.eventually.be.fulfilled;
+      });
     });
 
     it('fires \'destroy\' event', () => {
       let promise = new Promise((resolve) => {
         element.addEventListener('destroy', resolve);
       });
-      element.setActiveTab(0);
-      element.setActiveTab(1);
-      return expect(promise).to.eventually.be.fulfilled;
+
+      return element.setActiveTab(0).then(() => {
+        return element.loadPage('page1').then(() => {
+          element.setActiveTab(1);
+          return expect(promise).to.eventually.be.fulfilled;
+        });
+      });
     });
   });
 
@@ -310,10 +316,10 @@ describe('OnsTabbarElement', () => {
   });
 
   describe('#setActiveTab()', () => {
-    it('loads a persistent tab', (done) => {
+    it('loads any tab as persistent', (done) => {
       let element = ons._util.createElement(`
         <ons-tabbar>
-          <ons-tab label="Hoge" page="hoge" persistent></ons-tab>
+          <ons-tab label="Hoge" page="hoge"></ons-tab>
         </ons-tabbar>
       `);
       document.body.appendChild(element);
@@ -349,11 +355,11 @@ describe('OnsTabbarElement', () => {
       expect(element.setActiveTab(0)).to.be.false;
     });
 
-    it('does not remove persistent tabs', (done) => {
+    it('does not remove tabs', (done) => {
       let element = ons._util.createElement(`
         <ons-tabbar>
-          <ons-tab label="Hoge" page="hoge" persistent></ons-tab>
-          <ons-tab label="fuga" page="fuga" persistent></ons-tab>
+          <ons-tab label="Hoge" page="hoge"></ons-tab>
+          <ons-tab label="fuga" page="fuga"></ons-tab>
         </ons-tabbar>
       `);
 
@@ -371,8 +377,8 @@ describe('OnsTabbarElement', () => {
     it('keeps the page when option \'keepPage\' is true', (done) => {
       let element = ons._util.createElement(`
         <ons-tabbar>
-          <ons-tab label="Hoge" page="hoge" persistent active="true"></ons-tab>
-          <ons-tab label="fuga" page="fuga" persistent></ons-tab>
+          <ons-tab label="Hoge" page="hoge" active="true"></ons-tab>
+          <ons-tab label="fuga" page="fuga"></ons-tab>
         </ons-tabbar>
       `);